backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.2.1
openai==2.9.0
pydantic==1.10.12
